chore: remove resources calls
