+++ codersdk/workspaceagents.go
@@ -26,6 +26,61 @@ import (
 	"github.com/coder/retry"
 )
 
+type WorkspaceAgentStatus string
+
+const (
+	WorkspaceAgentConnecting   WorkspaceAgentStatus = "connecting"
+	WorkspaceAgentConnected    WorkspaceAgentStatus = "connected"
+	WorkspaceAgentDisconnected WorkspaceAgentStatus = "disconnected"
+)
+
+type WorkspaceAgent struct {
+	ID                   uuid.UUID            `json:"id"`
+	CreatedAt            time.Time            `json:"created_at"`
+	UpdatedAt            time.Time            `json:"updated_at"`
+	FirstConnectedAt     *time.Time           `json:"first_connected_at,omitempty"`
+	LastConnectedAt      *time.Time           `json:"last_connected_at,omitempty"`
+	DisconnectedAt       *time.Time           `json:"disconnected_at,omitempty"`
+	Status               WorkspaceAgentStatus `json:"status"`
+	Name                 string               `json:"name"`
+	ResourceID           uuid.UUID            `json:"resource_id"`
+	InstanceID           string               `json:"instance_id,omitempty"`
+	Architecture         string               `json:"architecture"`
+	EnvironmentVariables map[string]string    `json:"environment_variables"`
+	OperatingSystem      string               `json:"operating_system"`
+	StartupScript        string               `json:"startup_script,omitempty"`
+	Directory            string               `json:"directory,omitempty"`
+	Version              string               `json:"version"`
+	Apps                 []WorkspaceApp       `json:"apps"`
+	// DERPLatency is mapped by region name (e.g. "New York City", "Seattle").
+	DERPLatency map[string]DERPRegion `json:"latency,omitempty"`
+}
+
+type WorkspaceAgentResourceMetadata struct {
+	MemoryTotal uint64  `json:"memory_total"`
+	DiskTotal   uint64  `json:"disk_total"`
+	CPUCores    uint64  `json:"cpu_cores"`
+	CPUModel    string  `json:"cpu_model"`
+	CPUMhz      float64 `json:"cpu_mhz"`
+}
+
+type DERPRegion struct {
+	Preferred           bool    `json:"preferred"`
+	LatencyMilliseconds float64 `json:"latency_ms"`
+}
+
+type WorkspaceAgentInstanceMetadata struct {
+	JailOrchestrator   string `json:"jail_orchestrator"`
+	OperatingSystem    string `json:"operating_system"`
+	Platform           string `json:"platform"`
+	PlatformFamily     string `json:"platform_family"`
+	KernelVersion      string `json:"kernel_version"`
+	KernelArchitecture string `json:"kernel_architecture"`
+	Cloud              string `json:"cloud"`
+	Jail               string `json:"jail"`
+	VNC                bool   `json:"vnc"`
+}
+
 // @typescript-ignore GoogleInstanceIdentityToken
 type GoogleInstanceIdentityToken struct {
 	JSONWebToken string `json:"json_web_token" validate:"required"`
